This commit was manufactured by cvs2svn to create tag 'before-fedora-2_6_8-1_521-merge'.

+++ arch/parisc/lib/bitops.c
@@ -34,26 +34,26 @@ unsigned long __xchg64(unsigned long x, unsigned long *ptr)
 unsigned long __xchg32(int x, int *ptr)
 {
 	unsigned long flags;
-	long temp;
+	unsigned long temp;
 
 	atomic_spin_lock_irqsave(ATOMIC_HASH(ptr), flags);
-	temp = (long) *ptr;	/* XXX - sign extension wanted? */
+	(long) temp = (long) *ptr;	/* XXX - sign extension wanted? */
 	*ptr = x;
 	atomic_spin_unlock_irqrestore(ATOMIC_HASH(ptr), flags);
-	return (unsigned long)temp;
+	return temp;
 }
 
 
 unsigned long __xchg8(char x, char *ptr)
 {
 	unsigned long flags;
-	long temp;
+	unsigned long temp;
 
 	atomic_spin_lock_irqsave(ATOMIC_HASH(ptr), flags);
-	temp = (long) *ptr;	/* XXX - sign extension wanted? */
+	(long) temp = (long) *ptr;	/* XXX - sign extension wanted? */
 	*ptr = x;
 	atomic_spin_unlock_irqrestore(ATOMIC_HASH(ptr), flags);
-	return (unsigned long)temp;
+	return temp;
 }
